Move tests to dedicated module
