+++ actors/market/tests/market_actor_test.rs
@@ -1125,465 +1125,6 @@ fn assert_deal_failure<F>(
     // TODO: actor.checkState(rt)
 }
 
-#[cfg(test)]
-mod publish_storage_deals_failures {
-    use super::*;
-
-    #[test]
-    fn deal_end_after_deal_start() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.start_epoch = 10;
-            d.end_epoch = 9;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn current_epoch_greater_than_start_epoch() {
-        let f = |rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.start_epoch = rt.epoch - 1;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn deal_duration_greater_than_max_deal_duration() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.start_epoch = ChainEpoch::from(10);
-            d.end_epoch = d.start_epoch + (540 * EPOCHS_IN_DAY) + 1
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn negative_price_per_epoch() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.storage_price_per_epoch = TokenAmount::from(-1);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn price_per_epoch_greater_than_total_filecoin() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.storage_price_per_epoch = TOTAL_FILECOIN.clone() + 1;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn negative_provider_collateral() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.provider_collateral = TokenAmount::from(-1);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn provider_collateral_greater_than_max_collateral() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.provider_collateral = TOTAL_FILECOIN.clone() + 1;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn provider_collateral_less_than_bound() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            let power = StoragePower::from_i128(1 << 50).unwrap();
-            let (provider_min, _) = deal_provider_collateral_bounds(
-                &Policy::default(),
-                PaddedPieceSize(2048),
-                &BigInt::from(0u8),
-                &BigInt::from(0u8),
-                &power,
-            );
-            d.provider_collateral = provider_min - 1;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn negative_client_collateral() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.client_collateral = TokenAmount::from(-1);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn client_collateral_greater_than_max_collateral() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.client_collateral = TOTAL_FILECOIN.clone() + 1;
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn client_does_not_have_enough_balance_for_collateral() {
-        let f = |rt: &mut MockRuntime, d: &mut DealProposal| {
-            add_participant_funds(rt, CLIENT_ADDR, d.client_balance_requirement() - 1);
-            add_provider_funds(rt, d.provider_collateral.clone(), &MinerAddresses::default());
-        };
-        assert_deal_failure(false, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn provider_does_not_have_enough_balance_for_collateral() {
-        let f = |rt: &mut MockRuntime, d: &mut DealProposal| {
-            add_participant_funds(rt, CLIENT_ADDR, d.client_balance_requirement());
-            add_provider_funds(rt, d.provider_collateral.clone() - 1, &MinerAddresses::default());
-        };
-        assert_deal_failure(false, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn client_address_does_not_exist() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.client = Address::new_id(1);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn unable_to_resolve_client_address() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.client = new_bls_addr(1);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn signature_is_invalid() {
-        let f = |_rt: &mut MockRuntime, _d: &mut DealProposal| {};
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Err(anyhow!("error")));
-    }
-
-    #[test]
-    fn no_entry_for_client_in_locked_balance_table() {
-        let f = |rt: &mut MockRuntime, d: &mut DealProposal| {
-            add_provider_funds(rt, d.provider_collateral.clone(), &MinerAddresses::default());
-        };
-        assert_deal_failure(false, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn no_entry_for_provider_in_locked_balance_table() {
-        let f = |rt: &mut MockRuntime, d: &mut DealProposal| {
-            add_participant_funds(rt, CLIENT_ADDR, d.client_balance_requirement());
-        };
-        assert_deal_failure(false, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn bad_piece_cid() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.piece_cid = Cid::default();
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn zero_piece_size() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.piece_size = PaddedPieceSize(0u64);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn piece_size_less_than_128_bytes() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.piece_size = PaddedPieceSize(64u64);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn piece_size_is_not_a_power_of_2() {
-        let f = |_rt: &mut MockRuntime, d: &mut DealProposal| {
-            d.piece_size = PaddedPieceSize(254u64);
-        };
-        assert_deal_failure(true, f, ExitCode::USR_ILLEGAL_ARGUMENT, Ok(()));
-    }
-
-    #[test]
-    fn fail_when_client_has_some_funds_but_not_enough_for_a_deal() {
-        let mut rt = setup();
-
-        //
-        add_participant_funds(&mut rt, CLIENT_ADDR, TokenAmount::from(100u8));
-        let start_epoch = 42;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-        let deal1 = generate_deal_proposal(CLIENT_ADDR, PROVIDER_ADDR, start_epoch, end_epoch);
-        add_provider_funds(&mut rt, deal1.clone().provider_collateral, &MinerAddresses::default());
-        let buf = RawBytes::serialize(deal1.clone()).expect("failed to marshal deal proposal");
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal {
-                proposal: deal1.clone(),
-                client_signature: sig.clone(),
-            }],
-        };
-
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_provider_control_address(&mut rt, PROVIDER_ADDR, OWNER_ADDR, WORKER_ADDR);
-        expect_query_network_info(&mut rt);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        rt.expect_verify_signature(ExpectedVerifySig {
-            sig,
-            signer: deal1.client,
-            plaintext: buf.to_vec(),
-            result: Ok(()),
-        });
-        expect_abort(
-            ExitCode::USR_ILLEGAL_ARGUMENT,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-
-        rt.verify();
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fail_when_provider_has_some_funds_but_not_enough_for_a_deal() {
-        let start_epoch = 10;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-
-        let mut rt = setup();
-
-        add_provider_funds(&mut rt, TokenAmount::from(1u8), &MinerAddresses::default());
-        let deal1 = generate_deal_proposal(CLIENT_ADDR, PROVIDER_ADDR, start_epoch, end_epoch);
-        add_participant_funds(&mut rt, CLIENT_ADDR, deal1.client_balance_requirement());
-
-        let buf = RawBytes::serialize(deal1.clone()).expect("failed to marshal deal proposal");
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal {
-                proposal: deal1.clone(),
-                client_signature: sig.clone(),
-            }],
-        };
-
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_provider_control_address(&mut rt, PROVIDER_ADDR, OWNER_ADDR, WORKER_ADDR);
-        expect_query_network_info(&mut rt);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        rt.expect_verify_signature(ExpectedVerifySig {
-            sig,
-            signer: deal1.client,
-            plaintext: buf.to_vec(),
-            result: Ok(()),
-        });
-        expect_abort(
-            ExitCode::USR_ILLEGAL_ARGUMENT,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-
-        rt.verify();
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fail_when_deals_have_different_providers() {
-        let start_epoch = 10;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-
-        let mut rt = setup();
-        let deal1 = generate_deal_and_add_funds(
-            &mut rt,
-            CLIENT_ADDR,
-            &MinerAddresses::default(),
-            start_epoch,
-            end_epoch,
-        );
-        let m2 = MinerAddresses { provider: Address::new_id(1000), ..MinerAddresses::default() };
-
-        let deal2 = generate_deal_and_add_funds(&mut rt, CLIENT_ADDR, &m2, 1, end_epoch);
-
-        let buf1 = RawBytes::serialize(deal1.clone()).expect("failed to marshal deal proposal");
-        let buf2 = RawBytes::serialize(deal2.clone()).expect("failed to marshal deal proposal");
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![
-                ClientDealProposal { proposal: deal1.clone(), client_signature: sig.clone() },
-                ClientDealProposal { proposal: deal2.clone(), client_signature: sig.clone() },
-            ],
-        };
-
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_provider_control_address(&mut rt, PROVIDER_ADDR, OWNER_ADDR, WORKER_ADDR);
-        expect_query_network_info(&mut rt);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        rt.expect_verify_signature(ExpectedVerifySig {
-            sig: sig.clone(),
-            signer: deal1.client,
-            plaintext: buf1.to_vec(),
-            result: Ok(()),
-        });
-        rt.expect_verify_signature(ExpectedVerifySig {
-            sig,
-            signer: deal2.client,
-            plaintext: buf2.to_vec(),
-            result: Ok(()),
-        });
-
-        let psd_ret: PublishStorageDealsReturn = rt
-            .call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            )
-            .unwrap()
-            .deserialize()
-            .unwrap();
-
-        let valid: Vec<u64> = psd_ret.valid_deals.bounded_iter(std::u64::MAX).unwrap().collect();
-        assert_eq!(vec![0], valid);
-
-        rt.verify();
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fail_when_caller_is_not_of_signable_type() {
-        let start_epoch = 10;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-
-        let mut rt = setup();
-        let deal = generate_deal_proposal(CLIENT_ADDR, PROVIDER_ADDR, start_epoch, end_epoch);
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal { proposal: deal, client_signature: sig }],
-        };
-        let w = Address::new_id(1000);
-        rt.set_caller(*MINER_ACTOR_CODE_ID, w);
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_abort(
-            ExitCode::USR_FORBIDDEN,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fail_when_no_deals_in_params() {
-        let mut rt = setup();
-        let params = PublishStorageDealsParams { deals: vec![] };
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_abort(
-            ExitCode::USR_ILLEGAL_ARGUMENT,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fail_to_resolve_provider_address() {
-        let start_epoch = 10;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-
-        let mut rt = setup();
-        let mut deal = generate_deal_proposal(CLIENT_ADDR, PROVIDER_ADDR, start_epoch, end_epoch);
-        deal.provider = new_bls_addr(100);
-
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal { proposal: deal, client_signature: sig }],
-        };
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        expect_abort(
-            ExitCode::USR_NOT_FOUND,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-        check_state(&rt);
-    }
-
-    #[test]
-    fn caller_is_not_the_same_as_the_worker_address_for_miner() {
-        let start_epoch = 10;
-        let end_epoch = start_epoch + 200 * EPOCHS_IN_DAY;
-
-        let mut rt = setup();
-        let deal = generate_deal_proposal(CLIENT_ADDR, PROVIDER_ADDR, start_epoch, end_epoch);
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal { proposal: deal, client_signature: sig }],
-        };
-
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        let return_value = ext::miner::GetControlAddressesReturnParams {
-            owner: OWNER_ADDR,
-            worker: Address::new_id(999),
-            control_addresses: vec![],
-        };
-        rt.expect_send(
-            PROVIDER_ADDR,
-            ext::miner::CONTROL_ADDRESSES_METHOD,
-            RawBytes::default(),
-            TokenAmount::default(),
-            RawBytes::serialize(return_value).unwrap(),
-            ExitCode::OK,
-        );
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        expect_abort(
-            ExitCode::USR_FORBIDDEN,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-
-        rt.verify();
-        check_state(&rt);
-    }
-
-    #[test]
-    fn fails_if_provider_is_not_a_storage_miner_actor() {
-        let mut rt = setup();
-
-        // deal provider will be a Storage Miner Actor.
-        let p2 = Address::new_id(505);
-        rt.set_address_actor_type(p2, *POWER_ACTOR_CODE_ID);
-        let deal =
-            generate_deal_proposal(CLIENT_ADDR, p2, ChainEpoch::from(1), ChainEpoch::from(5));
-
-        let sig = Signature::new_bls("does not matter".as_bytes().to_vec());
-        let params = PublishStorageDealsParams {
-            deals: vec![ClientDealProposal { proposal: deal, client_signature: sig }],
-        };
-
-        rt.expect_validate_caller_type(vec![*ACCOUNT_ACTOR_CODE_ID, *MULTISIG_ACTOR_CODE_ID]);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, WORKER_ADDR);
-        expect_abort(
-            ExitCode::USR_ILLEGAL_ARGUMENT,
-            rt.call::<MarketActor>(
-                Method::PublishStorageDeals as u64,
-                &RawBytes::serialize(params).unwrap(),
-            ),
-        );
-
-        rt.verify();
-        check_state(&rt);
-    }
-}
-
 // Converted from: https://github.com/filecoin-project/specs-actors/blob/master/actors/builtin/market/market_test.go#L1519
 #[test]
 fn fail_when_deal_is_activated_but_proposal_is_not_found() {
